refactor: Content Search를 모달에서 mainContent 탭으로 변경
워크플로우 흐름을 방해하지 않도록 검색을 메인 뷰로 전환:
- ContentSearchModal 제거 → ContentSearchView 위젯으로 변경
- viewModeAtom에 'contentSearch' 타입 추가
- Cmd+Shift+F로 viewMode를 'contentSearch'로 전환
- contentSearchModalOpenAtom 제거 (viewMode로 표시/숨김 제어)
- IDE/CodeDoc처럼 메인 콘텐츠 영역에서 탭 전환 방식으로 동작
- ESC로 IDE 모드로 복귀
