+++ src/features/Search/ContentSearch/model/atoms.ts
@@ -1,13 +1,11 @@
 /**
  * ContentSearch Atoms
  * State management for content search feature
+ * Note: View visibility is controlled by viewModeAtom ('contentSearch')
  */
 
 import { atom } from 'jotai';
 import type { ContentSearchOptions, ContentSearchResult } from './types';
-
-// Modal open state
-export const contentSearchModalOpenAtom = atom<boolean>(false);
 
 // Search query
 export const contentSearchQueryAtom = atom<string>('');
